soc/mediatek/mt8196: Correct the argument type of MT6363
Correct the argument type of the mt6363_sdmadc_read API and the return
value type of the mt6363_read16 API.

TEST=Build pass
BUG=b:317009620

Reviewed-on: https://review.coreboot.org/c/coreboot/+/85839
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/mediatek/common/mt6363.c
@@ -25,7 +25,7 @@ static const struct pmic_setting key_protect_setting[] = {
 };
 
 static struct pmif *pmif_arb;
-u32 mt6363_read16(u32 reg)
+u16 mt6363_read16(u32 reg)
 {
 	u16 rdata = 0;
